requirements.txt
# ======================================
# PawFinders Pet Adoption System
# Python Dependencies
# ======================================

# Core Framework
flask==3.0.0
python-dotenv==1.0.0

# Machine Learning & AI
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.2
pillow==10.1.0
scikit-learn==1.3.2
pandas==2.1.4

# NLP & LangChain Dependencies
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.13
langchain-pinecone==0.0.3
langchain-openai==0.0.2
langchain-experimental==0.0.47
langchain-huggingface==0.0.1
langchain-groq==0.0.1
langchain-core==0.1.10

# Vector Database
pinecone-client[grpc]==3.0.0

# PDF Processing
pypdf==3.17.4
pypdf2==3.0.1

# HTTP & API
requests==2.31.0
urllib3==2.1.0

# Data Processing
python-multipart==0.0.6

# Development Tools
jupyter==1.0.0
ipython==8.18.1
notebook==7.0.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
colorama==0.4.6
